Move definitions for RPC_ADDR arg to rpc_addr
